feat(labels): add complete functionality for labels

+++ fe/src/pages/labels/labels.vue
@@ -78,6 +78,10 @@ const labelsLabel = {
         deleteLabel: function(id) {
             const self = this;
 
+            if (!confirm('Are you sure to delete?')) {
+                return false;
+            }
+
             this.formLoading = true;
 
             axios.delete(`/api/labels/delete/${id}`)
